Use select() instead of selectAll() where applicable.
select() is much more resource friendly, because it doesn't copy the complete result set into an array, but uses the database's result paging capabilities instead while traversing over the result set.
It can safely be used where we simply loop over the result set. If we need the array structure instead, or want to manipulate the rows inside the loop, selectAll() should still be used.